feat: add error handler

+++ src/command.ts
@@ -18,15 +18,19 @@ export abstract class Command {
 
 export class UpCommand extends Command {
   execute() {
-    this.currentPath.path = path.basename(this.currentPath.path);
-    console.log(`You are in "${process.cwd()}"`);
+    let path = process.cwd().split("\\");
+    let newPath = path.slice(0, path.length - 1).join("\\");
+    this.currentPath.path = newPath;
   }
 }
 
 export class CdCommand extends Command {
   execute() {
-    this.currentPath.path = this.params[0];
-    console.log(`You are in "${process.cwd()}"`);
+    if (fs.existsSync(this.params[0])) {
+      this.currentPath.path = this.params[0];
+    } else {
+      throw new Error(`Path: ${this.params[0]} doesn't exist!`);
+    }
   }
 }
 
@@ -44,7 +48,7 @@ export class LsCommand extends Command {
 export class CatCommand extends Command {
   execute() {
     const filePath = this.params[0];
-    const readableStream = fs.createReadStream(filePath);
+    const readableStream = fs.createReadStream(filePath, "utf-8");
     readableStream.on("error", function (error) {
       console.log(`error: ${error.message}`);
     });
@@ -56,21 +60,11 @@ export class CatCommand extends Command {
 
 export class AddCommand extends Command {
   execute() {
-    const __filename = fileURLToPath(import.meta.url);
-    const __dirname = dirname(__filename);
-    const filePath = path.join(__dirname, this.params[0]);
-    // fs.writeFile(filePath, "", (err) => {
-    //   if (err) {
-    //     console.log("Add operation failed due to: ", err);
-    //   } else {
-    //     console.log("New file added  successfully!");
-    //   }
-    // });
-
+    const filePath = path.join(this.currentPath.path, this.params[0]);
     let writeStream = fs.createWriteStream(filePath);
     writeStream.write("", "base64");
     writeStream.on("finish", () => {
-      console.log("wrote all data to file");
+      console.log("New file added  successfully!");
     });
     writeStream.end();
   }
@@ -82,10 +76,14 @@ export class RnCommand extends Command {
       path.basename(this.currentPath.path),
       this.params[1]
     );
-    fs.rename(this.params[0], newFileName, (err) => {
-      if (err) throw Error("Rn operation failed");
-      console.log("File renamed successfully");
-    });
+    if (fs.existsSync(this.params[0])) {
+      fs.rename(this.params[0], newFileName, (err) => {
+        if (err) throw Error("Rn operation failed");
+        console.log("File renamed successfully");
+      });
+    } else {
+      throw new Error(`Path to file doesn't correct!`);
+    }
   }
 }
 
@@ -93,10 +91,14 @@ export class CpCommand extends Command {
   execute() {
     const to = this.params[1];
     const from = this.params[0];
-    fs.copyFile(from, to, (err) => {
-      if (err) throw Error("Cp operation failed");
-      console.log("File copied successfully");
-    });
+    if (fs.existsSync(this.params[0])) {
+      fs.copyFile(from, to, (err) => {
+        if (err) throw Error("Cp operation failed");
+        console.log("File copied successfully");
+      });
+    } else {
+      throw new Error(`Path to file doesn't correct!`);
+    }
   }
 }
 
@@ -104,22 +106,30 @@ export class MvCommand extends Command {
   execute() {
     const to = this.params[1];
     const from = this.params[0];
-    fs.copyFile(from, to, (err) => {
-      if (err) throw Error("Mv operation failed");
-      console.log("File moved successfully");
-    });
-    fs.unlink(from, (err) => {
-      if (err) throw Error("Mv operation failed");
-    });
+    if (fs.existsSync(this.params[0]) || fs.existsSync(this.params[0])) {
+      fs.copyFile(from, to, (err) => {
+        if (err) throw Error(`Mv operation failed: ${err}`);
+        fs.unlink(from, (err) => {
+          if (err) throw Error(`Mv operation failed: ${err}`);
+          console.log("File moved successfully");
+        });
+      });
+    } else {
+      throw new Error(`Path doesn't correct!`);
+    }
   }
 }
 
 export class RmCommand extends Command {
   execute() {
+    if(fs.existsSync(this.params[0])){
     fs.unlink(this.params[0], (err) => {
       if (err) throw Error("Rm operation failed");
       console.log("File deleted successfully");
     });
+  } else {
+    throw new Error(`Path doesn't correct!`);
+  }
   }
 }
 
@@ -127,7 +137,7 @@ export class OsCommand extends Command {
   execute() {
     switch (this.params[0]) {
       case "--EOL":
-        console.log(os.EOL);
+        console.log(`End-Of-Line: ${JSON.stringify(os.EOL)}`);
         break;
       case "--cpus":
         console.log(os.cpus());
@@ -147,6 +157,7 @@ export class OsCommand extends Command {
 
 export class HashCommand extends Command {
   execute() {
+    if(fs.existsSync(this.params[0])){
     const readableStream = fs.createReadStream(this.params[0]);
     readableStream.on("error", function (error) {
       console.log(`error: ${error.message}`);
@@ -155,41 +166,48 @@ export class HashCommand extends Command {
       let hash = crypto.createHash("sha256").update(chunk).digest("hex");
       console.log(hash);
     });
-
-    // fs.readFile(this.params[0], "utf-8", (err, data) => {
-    //   if (err) throw Error("Hash operation failed");
-    //   let hash = crypto.createHash("sha256").update(data).digest("hex");
-    //   console.log(hash);
-    // });
+  } else {
+    throw new Error(`Path to file doesn't correct!`);
   }
+}
 }
 
 export class CompressCommand extends Command {
   execute() {
     const inpPath = this.params[0];
     const outPath = this.params[1];
+    if(fs.existsSync(this.params[0])){
     const gzip = zlib.createGzip();
     const inp = fs.createReadStream(inpPath);
     var out = fs.createWriteStream(outPath);
     inp.pipe(gzip).pipe(out);
+  }else {
+    throw new Error(`Path to file doesn't correct!`);
   }
+} 
 }
 
 export class DecompressCommand extends Command {
   execute() {
-    const decompressedPath = this.params[0];
-    const compressedPath = this.params[1];
+    const decompressedPath = this.params[1];
+    const compressedPath = this.params[0];
+    if(fs.existsSync(this.params[0])){
     var unzip = zlib.createUnzip();
     var read = fs.createReadStream(compressedPath);
     var write = fs.createWriteStream(decompressedPath);
     read.pipe(unzip).pipe(write);
     console.log("unZipped Successfully");
+  }else {
+    throw new Error(`Path to file doesn't correct!`);
   }
+}
 }
 
 export class ExitCommand extends Command {
   execute(): void {
+    const args = process.argv.slice(2);
+    const username = args[0].split("=")[1];
+    console.log(`\n Thank you for using File Manager, ${username}`);
     process.exit();
-  }  
+  }
 }
-
